Share instead of duplicate options code
