--- src/CodeStyle/Core/Analyzers/Options/OptionStorageLocation.cs
@@ -1,11 +1,0 @@
-﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
-
-namespace Microsoft.CodeAnalysis.Options
-{
-    /// <summary>
-    /// The base type of all types that specify where options are stored.
-    /// </summary>
-    internal abstract class OptionStorageLocation
-    {
-    }
-}
