+++ src/CodeStyle/Core/Analyzers/PublicAPI.Unshipped.txt
@@ -1,0 +1,55 @@
+Microsoft.CodeAnalysis.CodeStyle.CodeStyleOption<T>
+Microsoft.CodeAnalysis.CodeStyle.CodeStyleOption<T>.CodeStyleOption(T value, Microsoft.CodeAnalysis.CodeStyle.NotificationOption notification) -> void
+Microsoft.CodeAnalysis.CodeStyle.CodeStyleOption<T>.Equals(Microsoft.CodeAnalysis.CodeStyle.CodeStyleOption<T> other) -> bool
+Microsoft.CodeAnalysis.CodeStyle.CodeStyleOption<T>.Notification.get -> Microsoft.CodeAnalysis.CodeStyle.NotificationOption
+Microsoft.CodeAnalysis.CodeStyle.CodeStyleOption<T>.Notification.set -> void
+Microsoft.CodeAnalysis.CodeStyle.CodeStyleOption<T>.ToXElement() -> System.Xml.Linq.XElement
+Microsoft.CodeAnalysis.CodeStyle.CodeStyleOption<T>.Value.get -> T
+Microsoft.CodeAnalysis.CodeStyle.CodeStyleOption<T>.Value.set -> void
+Microsoft.CodeAnalysis.CodeStyle.NotificationOption
+Microsoft.CodeAnalysis.CodeStyle.NotificationOption.Name.get -> string
+Microsoft.CodeAnalysis.CodeStyle.NotificationOption.Name.set -> void
+Microsoft.CodeAnalysis.CodeStyle.NotificationOption.Severity.get -> Microsoft.CodeAnalysis.ReportDiagnostic
+Microsoft.CodeAnalysis.CodeStyle.NotificationOption.Severity.set -> void
+Microsoft.CodeAnalysis.CodeStyle.NotificationOption.Value.get -> Microsoft.CodeAnalysis.DiagnosticSeverity
+Microsoft.CodeAnalysis.CodeStyle.NotificationOption.Value.set -> void
+Microsoft.CodeAnalysis.Options.IOption
+Microsoft.CodeAnalysis.Options.IOption.DefaultValue.get -> object
+Microsoft.CodeAnalysis.Options.IOption.Feature.get -> string
+Microsoft.CodeAnalysis.Options.IOption.IsPerLanguage.get -> bool
+Microsoft.CodeAnalysis.Options.IOption.Name.get -> string
+Microsoft.CodeAnalysis.Options.IOption.StorageLocations.get -> System.Collections.Immutable.ImmutableArray<Microsoft.CodeAnalysis.Options.OptionStorageLocation>
+Microsoft.CodeAnalysis.Options.IOption.Type.get -> System.Type
+Microsoft.CodeAnalysis.Options.Option<T>
+Microsoft.CodeAnalysis.Options.Option<T>.DefaultValue.get -> T
+Microsoft.CodeAnalysis.Options.Option<T>.Feature.get -> string
+Microsoft.CodeAnalysis.Options.Option<T>.Name.get -> string
+Microsoft.CodeAnalysis.Options.Option<T>.Option(string feature, string name) -> void
+Microsoft.CodeAnalysis.Options.Option<T>.Option(string feature, string name, T defaultValue) -> void
+Microsoft.CodeAnalysis.Options.Option<T>.Option(string feature, string name, T defaultValue, params Microsoft.CodeAnalysis.Options.OptionStorageLocation[] storageLocations) -> void
+Microsoft.CodeAnalysis.Options.Option<T>.StorageLocations.get -> System.Collections.Immutable.ImmutableArray<Microsoft.CodeAnalysis.Options.OptionStorageLocation>
+Microsoft.CodeAnalysis.Options.Option<T>.Type.get -> System.Type
+Microsoft.CodeAnalysis.Options.OptionKey
+Microsoft.CodeAnalysis.Options.OptionKey.Equals(Microsoft.CodeAnalysis.Options.OptionKey other) -> bool
+Microsoft.CodeAnalysis.Options.OptionKey.Language.get -> string
+Microsoft.CodeAnalysis.Options.OptionKey.Option.get -> Microsoft.CodeAnalysis.Options.IOption
+Microsoft.CodeAnalysis.Options.OptionKey.OptionKey(Microsoft.CodeAnalysis.Options.IOption option, string language = null) -> void
+Microsoft.CodeAnalysis.Options.OptionStorageLocation
+Microsoft.CodeAnalysis.Options.OptionStorageLocation.OptionStorageLocation() -> void
+override Microsoft.CodeAnalysis.CodeStyle.CodeStyleOption<T>.Equals(object obj) -> bool
+override Microsoft.CodeAnalysis.CodeStyle.CodeStyleOption<T>.GetHashCode() -> int
+override Microsoft.CodeAnalysis.CodeStyle.NotificationOption.ToString() -> string
+override Microsoft.CodeAnalysis.Options.Option<T>.ToString() -> string
+override Microsoft.CodeAnalysis.Options.OptionKey.Equals(object obj) -> bool
+override Microsoft.CodeAnalysis.Options.OptionKey.GetHashCode() -> int
+override Microsoft.CodeAnalysis.Options.OptionKey.ToString() -> string
+static Microsoft.CodeAnalysis.CodeStyle.CodeStyleOption<T>.Default.get -> Microsoft.CodeAnalysis.CodeStyle.CodeStyleOption<T>
+static Microsoft.CodeAnalysis.CodeStyle.CodeStyleOption<T>.FromXElement(System.Xml.Linq.XElement element) -> Microsoft.CodeAnalysis.CodeStyle.CodeStyleOption<T>
+static Microsoft.CodeAnalysis.Options.Option<T>.implicit operator Microsoft.CodeAnalysis.Options.OptionKey(Microsoft.CodeAnalysis.Options.Option<T> option) -> Microsoft.CodeAnalysis.Options.OptionKey
+static Microsoft.CodeAnalysis.Options.OptionKey.operator !=(Microsoft.CodeAnalysis.Options.OptionKey left, Microsoft.CodeAnalysis.Options.OptionKey right) -> bool
+static Microsoft.CodeAnalysis.Options.OptionKey.operator ==(Microsoft.CodeAnalysis.Options.OptionKey left, Microsoft.CodeAnalysis.Options.OptionKey right) -> bool
+static readonly Microsoft.CodeAnalysis.CodeStyle.NotificationOption.Error -> Microsoft.CodeAnalysis.CodeStyle.NotificationOption
+static readonly Microsoft.CodeAnalysis.CodeStyle.NotificationOption.None -> Microsoft.CodeAnalysis.CodeStyle.NotificationOption
+static readonly Microsoft.CodeAnalysis.CodeStyle.NotificationOption.Silent -> Microsoft.CodeAnalysis.CodeStyle.NotificationOption
+static readonly Microsoft.CodeAnalysis.CodeStyle.NotificationOption.Suggestion -> Microsoft.CodeAnalysis.CodeStyle.NotificationOption
+static readonly Microsoft.CodeAnalysis.CodeStyle.NotificationOption.Warning -> Microsoft.CodeAnalysis.CodeStyle.NotificationOption
