+++ src/CodeStyle/Core/Analyzers/xlf/CodeStyleResources.it.xlf
@@ -12,14 +12,39 @@
         <target state="new">A language name must be specified for this option.</target>
         <note />
       </trans-unit>
+      <trans-unit id="Error">
+        <source>Error</source>
+        <target state="new">Error</target>
+        <note />
+      </trans-unit>
       <trans-unit id="Fix_formatting">
         <source>Fix formatting</source>
         <target state="new">Fix formatting</target>
         <note />
       </trans-unit>
+      <trans-unit id="None">
+        <source>None</source>
+        <target state="new">None</target>
+        <note />
+      </trans-unit>
+      <trans-unit id="Refactoring_Only">
+        <source>Refactoring Only</source>
+        <target state="new">Refactoring Only</target>
+        <note />
+      </trans-unit>
       <trans-unit id="Style">
         <source>Style</source>
         <target state="new">Style</target>
+        <note />
+      </trans-unit>
+      <trans-unit id="Suggestion">
+        <source>Suggestion</source>
+        <target state="new">Suggestion</target>
+        <note />
+      </trans-unit>
+      <trans-unit id="Warning">
+        <source>Warning</source>
+        <target state="new">Warning</target>
         <note />
       </trans-unit>
     </body>
