+++ src/Workspaces/Core/Portable/CodeStyle/NotificationOption.cs
@@ -3,6 +3,10 @@
 using System;
 using Microsoft.CodeAnalysis.Diagnostics;
 
+#if CODE_STYLE
+using WorkspacesResources = Microsoft.CodeAnalysis.CodeStyleResources;
+#endif
+
 namespace Microsoft.CodeAnalysis.CodeStyle
 {
     /// <summary>
